requirements.txt
ar_wordcloud==0.0.4
arabic_reshaper==3.0.0
matplotlib==3.7.1
nltk==3.8.1
numpy==1.24.3
pandas==1.5.3
Pillow==9.4.0
Pillow==11.0.0
Requests==2.32.3
scikit_learn==1.3.0
scikit_learn==1.3.2
streamlit==1.22.0
streamlit==1.24.1
streamlit==1.29.0
streamlit_lottie==0.0.5
streamlit_option_menu==0.3.6
tensorflow==2.18.0
tensorflow_intel==2.18.0
wordcloud==1.9.3
